+++ httpx/_transports/default.py
@@ -49,7 +49,7 @@ from .._exceptions import (
     WriteTimeout,
 )
 from .._types import CertTypes, VerifyTypes
-from .base import AsyncBaseTransport, BaseTransport
+from .base import AsyncBaseTransport, AsyncByteStream, BaseTransport, SyncByteStream
 
 T = typing.TypeVar("T", bound="HTTPTransport")
 A = typing.TypeVar("A", bound="AsyncHTTPTransport")
@@ -78,19 +78,6 @@ def map_httpcore_exceptions() -> typing.Iterator[None]:
         raise mapped_exc(message) from exc
 
 
-def ensure_http_version_reason_phrase_as_bytes(extensions: dict) -> None:
-    # From HTTPX 0.18 onwards we're treating the "reason_phrase" and "http_version"
-    # extensions as bytes, in order to be more precise. Also we're using the
-    # "reason_phrase" key in preference to "reason", in order to match properly
-    # with the HTTP spec naming.
-    # HTTPCore 0.12 does not yet use these same conventions for the extensions,
-    # so we bridge between the two styles for now.
-    if "reason" in extensions:
-        extensions["reason_phrase"] = extensions.pop("reason").encode("ascii")
-    if "http_version" in extensions:
-        extensions["http_version"] = extensions["http_version"].encode("ascii")
-
-
 HTTPCORE_EXC_MAP = {
     httpcore.TimeoutException: TimeoutException,
     httpcore.ConnectTimeout: ConnectTimeout,
@@ -108,6 +95,20 @@ HTTPCORE_EXC_MAP = {
     httpcore.LocalProtocolError: LocalProtocolError,
     httpcore.RemoteProtocolError: RemoteProtocolError,
 }
+
+
+class ResponseStream(SyncByteStream):
+    def __init__(self, httpcore_stream: httpcore.SyncByteStream):
+        self._httpcore_stream = httpcore_stream
+
+    def __iter__(self) -> typing.Iterator[bytes]:
+        with map_httpcore_exceptions():
+            for part in self._httpcore_stream:
+                yield part
+
+    def close(self) -> None:
+        with map_httpcore_exceptions():
+            self._httpcore_stream.close()
 
 
 class HTTPTransport(BaseTransport):
@@ -161,43 +162,48 @@ class HTTPTransport(BaseTransport):
         exc_value: BaseException = None,
         traceback: TracebackType = None,
     ) -> None:
-        self._pool.__exit__(exc_type, exc_value, traceback)
+        with map_httpcore_exceptions():
+            self._pool.__exit__(exc_type, exc_value, traceback)
 
     def handle_request(
         self,
         method: bytes,
         url: typing.Tuple[bytes, bytes, typing.Optional[int], bytes],
         headers: typing.List[typing.Tuple[bytes, bytes]],
-        stream: typing.Iterable[bytes],
+        stream: SyncByteStream,
         extensions: dict,
     ) -> typing.Tuple[
-        int, typing.List[typing.Tuple[bytes, bytes]], typing.Iterable[bytes], dict
+        int, typing.List[typing.Tuple[bytes, bytes]], SyncByteStream, dict
     ]:
         with map_httpcore_exceptions():
-            status_code, headers, byte_stream, extensions = self._pool.request(
+            status_code, headers, byte_stream, extensions = self._pool.handle_request(
                 method=method,
                 url=url,
                 headers=headers,
                 stream=httpcore.IteratorByteStream(iter(stream)),
-                ext=extensions,
+                extensions=extensions,
             )
 
-        def response_stream() -> typing.Iterator[bytes]:
-            with map_httpcore_exceptions():
-                for part in byte_stream:
-                    yield part
+        stream = ResponseStream(byte_stream)
 
-        def close() -> None:
-            with map_httpcore_exceptions():
-                byte_stream.close()
-
-        ensure_http_version_reason_phrase_as_bytes(extensions)
-        extensions["close"] = close
-
-        return status_code, headers, response_stream(), extensions
+        return status_code, headers, stream, extensions
 
     def close(self) -> None:
         self._pool.close()
+
+
+class AsyncResponseStream(AsyncByteStream):
+    def __init__(self, httpcore_stream: httpcore.AsyncByteStream):
+        self._httpcore_stream = httpcore_stream
+
+    async def __aiter__(self) -> typing.AsyncIterator[bytes]:
+        with map_httpcore_exceptions():
+            async for part in self._httpcore_stream:
+                yield part
+
+    async def aclose(self) -> None:
+        with map_httpcore_exceptions():
+            await self._httpcore_stream.aclose()
 
 
 class AsyncHTTPTransport(AsyncBaseTransport):
@@ -251,40 +257,36 @@ class AsyncHTTPTransport(AsyncBaseTransport):
         exc_value: BaseException = None,
         traceback: TracebackType = None,
     ) -> None:
-        await self._pool.__aexit__(exc_type, exc_value, traceback)
+        with map_httpcore_exceptions():
+            await self._pool.__aexit__(exc_type, exc_value, traceback)
 
     async def handle_async_request(
         self,
         method: bytes,
         url: typing.Tuple[bytes, bytes, typing.Optional[int], bytes],
         headers: typing.List[typing.Tuple[bytes, bytes]],
-        stream: typing.AsyncIterable[bytes],
+        stream: AsyncByteStream,
         extensions: dict,
     ) -> typing.Tuple[
-        int, typing.List[typing.Tuple[bytes, bytes]], typing.AsyncIterable[bytes], dict
+        int, typing.List[typing.Tuple[bytes, bytes]], AsyncByteStream, dict
     ]:
         with map_httpcore_exceptions():
-            status_code, headers, byte_stream, extensions = await self._pool.arequest(
+            (
+                status_code,
+                headers,
+                byte_stream,
+                extensions,
+            ) = await self._pool.handle_async_request(
                 method=method,
                 url=url,
                 headers=headers,
                 stream=httpcore.AsyncIteratorByteStream(stream.__aiter__()),
-                ext=extensions,
+                extensions=extensions,
             )
 
-        async def response_stream() -> typing.AsyncIterator[bytes]:
-            with map_httpcore_exceptions():
-                async for part in byte_stream:
-                    yield part
+        stream = AsyncResponseStream(byte_stream)
 
-        async def aclose() -> None:
-            with map_httpcore_exceptions():
-                await byte_stream.aclose()
-
-        ensure_http_version_reason_phrase_as_bytes(extensions)
-        extensions["aclose"] = aclose
-
-        return status_code, headers, response_stream(), extensions
+        return status_code, headers, stream, extensions
 
     async def aclose(self) -> None:
         await self._pool.aclose()